fix: prevent duplicate js plugin registration
